+++ test/prerna/auth/utils/SecurityAPIUserUtilsUnitTests.java
@@ -69,7 +69,7 @@ public class SecurityAPIUserUtilsUnitTests extends AbstractSecurityUtilsUnitTest
         boolean exporter = false;
         String modelUsageRestriction = null;
         String modelUsageFrequency = null;
-        Integer modelMaxTokens = null;
+        Long modelMaxTokens = null;
         Double modelMaxResponseTime = null;
 
         boolean success = SecurityUpdateUtils.registerUser(id, name, email, password, type, phone, phoneextension,
