+++ test/prerna/auth/utils/SecurityShareSessionUtilsUnitTests.java
@@ -233,7 +233,7 @@ public class SecurityShareSessionUtilsUnitTests extends AbstractSecurityUtilsUni
 		boolean exporter = false;
 		String modelUsageRestriction = null;
 		String modelUsageFrequency = null;
-		Integer modelMaxTokens = null;
+		Long modelMaxTokens = null;
 		Double modelMaxResponseTime = null;
 
 		boolean success = SecurityUpdateUtils.registerUser(id, name, email, password, type, phone, phoneextension,
